feat: react native bare linking shared UI (#188)
* removing old components

* Adding AppNavigator

* Using shared ui

* Updating prettier

* removing initial App.tsx
