+++ examples/bare/.prettierrc.js
@@ -1,7 +1,3 @@
 module.exports = {
-  arrowParens: 'avoid',
-  bracketSameLine: true,
-  bracketSpacing: false,
-  singleQuote: true,
-  trailingComma: 'all',
+  extends: '../../.prettierrc.js',
 };
