+++ examples/bare/index.js
@@ -1,7 +1,7 @@
 import 'react-native-gesture-handler';
 
 import {AppRegistry} from 'react-native';
-import App from './App';
+import App from './src/App';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
